# docs/API_OpenAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "local-judge",
    "description": "Local native v1 evaluation, replay, and Jev adapter routes.",
    "version": "v1"
  },
  "paths": {
    "/v1/evaluations": {
      "post": {
        "summary": "Evaluations",
        "operationId": "evaluations_v1_evaluations_post",
        "responses": {
          "200": {
            "description": "native completed result",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Evaluations V1 Evaluations Post"
                }
              }
            }
          },
          "400": {
            "description": "native rejected result, or REQUEST_TOO_LARGE caused by more than 64 questions"
          },
          "413": {
            "description": "REQUEST_TOO_LARGE caused by the 256 KiB encoded-size limit"
          }
        }
      }
    },
    "/v1/replays": {
      "post": {
        "summary": "Replays",
        "operationId": "replays_v1_replays_post",
        "responses": {
          "200": {
            "description": "native completed result",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Replays V1 Replays Post"
                }
              }
            }
          },
          "400": {
            "description": "invalid replay body"
          },
          "409": {
            "description": "REPLAY_CONFIGURATION_UNAVAILABLE"
          }
        }
      }
    },
    "/v1/jev/evaluations": {
      "post": {
        "summary": "Jev Evaluations",
        "operationId": "jev_evaluations_v1_jev_evaluations_post",
        "responses": {
          "200": {
            "description": "Jev adapter result",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Jev Evaluations V1 Jev Evaluations Post"
                }
              }
            }
          },
          "400": {
            "description": "adapter result whose error is a native structural code (input validation) or JEV_ADAPTER_UNMAPPABLE_RESULT (unmappable result)"
          }
        }
      }
    }
  },
...
}